chore: create useReadyStateListeners composable
